fix c++/cuda code for multi-platform building

+++ torch_ans/rans.cpp
@@ -24,14 +24,14 @@ torch::Tensor rans_byte_strings_to_stream(std::vector<py::bytes> byte_strings)
 {
   // NOTE: implicit convert py::bytes to std::string
   // const std::string max_length_bytes = *std::max_element(byte_strings.begin(), byte_strings.end(), [](std::string s1, std::string s2){return s1.size()<s2.size();});
-  const ssize_t batch_size = byte_strings.size();
+  const int64_t batch_size = byte_strings.size();
   size_t max_length = 0;
   for (std::string bytes : byte_strings)
   {
     if (bytes.size() > max_length) max_length = bytes.size();
   }
 
-  const ssize_t stream_length = (max_length + sizeof(DEFAULT_TORCH_TENSOR_TYPE) - 1) / sizeof(DEFAULT_TORCH_TENSOR_TYPE) + 1;
+  const int64_t stream_length = (max_length + sizeof(DEFAULT_TORCH_TENSOR_TYPE) - 1) / sizeof(DEFAULT_TORCH_TENSOR_TYPE) + 1;
 
   auto stream = torch::zeros({batch_size, stream_length}, torch::TensorOptions().dtype(DEFAULT_TORCH_TENSOR_DTYPE));
   auto stream_accessor = stream.accessor<DEFAULT_TORCH_TENSOR_TYPE, 2>();
@@ -55,17 +55,17 @@ torch::Tensor rans_byte_strings_to_stream(std::vector<py::bytes> byte_strings)
 
 std::tuple<torch::Tensor, torch::Tensor> rans_alias_build_table(
   const torch::Tensor& cdfs, const torch::Tensor& cdfs_sizes,
-  ssize_t symbol_precision,
-  ssize_t freq_precision
+  int64_t symbol_precision,
+  int64_t freq_precision
 )
 {
   auto cdfs_accessor = cdfs.accessor<DEFAULT_TORCH_TENSOR_TYPE, 2>();
   auto cdfs_sizes_accessor = cdfs_sizes.accessor<DEFAULT_TORCH_TENSOR_TYPE, 1>();
-  const ssize_t batch_size = cdfs.size(0);
-  const ssize_t max_cdf_size = cdfs.size(1);
+  const int64_t batch_size = cdfs.size(0);
+  const int64_t max_cdf_size = cdfs.size(1);
 
   // TODO: confirm the max size of cdfs_with_alias_table
-  auto cdfs_with_alias_table = torch::zeros({batch_size, static_cast<ssize_t>(max_cdf_size + max_cdf_size*(sizeof(RANSAliasSamplingCDFTableElement<DEFAULT_TORCH_TENSOR_TYPE>)/sizeof(DEFAULT_TORCH_TENSOR_TYPE)))}, 
+  auto cdfs_with_alias_table = torch::zeros({batch_size, static_cast<int64_t>(max_cdf_size + max_cdf_size*(sizeof(RANSAliasSamplingCDFTableElement<DEFAULT_TORCH_TENSOR_TYPE>)/sizeof(DEFAULT_TORCH_TENSOR_TYPE)))}, 
     torch::TensorOptions().dtype(DEFAULT_TORCH_TENSOR_DTYPE));
   auto cdfs_with_alias_table_accessor = cdfs_with_alias_table.accessor<DEFAULT_TORCH_TENSOR_TYPE, 2>();
   auto cdfs_with_alias_remap = torch::zeros({batch_size, max_cdf_size + (1<<freq_precision)}, torch::TensorOptions().dtype(DEFAULT_TORCH_TENSOR_DTYPE));
